+++ packages/eslint-plugin/src/util/astUtils.ts
@@ -8,7 +8,7 @@ const LINEBREAK_MATCHER = /\r\n|[\r\n\u2028\u2029]/;
 
 function isOptionalChainPunctuator(
   token: TSESTree.Token | TSESTree.Comment,
-): boolean {
+): token is TSESTree.PunctuatorToken & { value: '?.' } {
   return token.type === AST_TOKEN_TYPES.Punctuator && token.value === '?.';
 }
 function isNotOptionalChainPunctuator(
@@ -19,7 +19,7 @@ function isNotOptionalChainPunctuator(
 
 function isNonNullAssertionPunctuator(
   token: TSESTree.Token | TSESTree.Comment,
-): boolean {
+): token is TSESTree.PunctuatorToken & { value: '!' } {
   return token.type === AST_TOKEN_TYPES.Punctuator && token.value === '!';
 }
 function isNotNonNullAssertionPunctuator(
@@ -33,7 +33,7 @@ function isNotNonNullAssertionPunctuator(
  */
 function isOptionalOptionalChain(
   node: TSESTree.Node,
-): node is TSESTree.OptionalCallExpression {
+): node is TSESTree.OptionalCallExpression & { optional: true } {
   return (
     node.type === AST_NODE_TYPES.OptionalCallExpression &&
     // this flag means the call expression itself is option
@@ -45,7 +45,9 @@ function isOptionalOptionalChain(
 /**
  * Returns true if and only if the node represents logical OR
  */
-function isLogicalOrOperator(node: TSESTree.Node): boolean {
+function isLogicalOrOperator(
+  node: TSESTree.Node,
+): node is TSESTree.LogicalExpression & { operator: '||' } {
   return (
     node.type === AST_NODE_TYPES.LogicalExpression && node.operator === '||'
   );
@@ -55,16 +57,18 @@ function isLogicalOrOperator(node: TSESTree.Node): boolean {
  * Determines whether two adjacent tokens are on the same line
  */
 function isTokenOnSameLine(
-  left: TSESTree.Token,
-  right: TSESTree.Token,
+  left: TSESTree.Token | TSESTree.Comment,
+  right: TSESTree.Token | TSESTree.Comment,
 ): boolean {
   return left.loc.end.line === right.loc.start.line;
 }
 
 /**
  * Checks if a node is a type assertion:
- * - x as foo
- * - <foo>x
+ * ```
+ * x as foo
+ * <foo>x
+ * ```
  */
 function isTypeAssertion(
   node: TSESTree.Node | undefined | null,
@@ -78,14 +82,49 @@ function isTypeAssertion(
   );
 }
 
+/**
+ * Checks if a node is a constructor method.
+ */
+function isConstructor(
+  node: TSESTree.Node | undefined,
+): node is TSESTree.MethodDefinition {
+  return (
+    node?.type === AST_NODE_TYPES.MethodDefinition &&
+    node.kind === 'constructor'
+  );
+}
+
+/**
+ * Checks if a node is a setter method.
+ */
+function isSetter(
+  node: TSESTree.Node | undefined,
+): node is TSESTree.MethodDefinition | TSESTree.Property {
+  return (
+    !!node &&
+    (node.type === AST_NODE_TYPES.MethodDefinition ||
+      node.type === AST_NODE_TYPES.Property) &&
+    node.kind === 'set'
+  );
+}
+
+function isIdentifier(
+  node: TSESTree.Node | undefined,
+): node is TSESTree.Identifier {
+  return node?.type === AST_NODE_TYPES.Identifier;
+}
+
 export {
-  isTypeAssertion,
+  isConstructor,
+  isIdentifier,
+  isLogicalOrOperator,
   isNonNullAssertionPunctuator,
   isNotNonNullAssertionPunctuator,
   isNotOptionalChainPunctuator,
   isOptionalChainPunctuator,
   isOptionalOptionalChain,
+  isSetter,
   isTokenOnSameLine,
-  isLogicalOrOperator,
+  isTypeAssertion,
   LINEBREAK_MATCHER,
 };