+++ packages/eslint-plugin/tests/rules/no-unused-vars-experimental.test.ts
@@ -543,6 +543,7 @@ export class Foo {
           messageId: 'unusedWithIgnorePattern',
           data: {
             name: 'foo',
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             type: 'Property',
             pattern: DEFAULT_IGNORED_REGEX,
           },
@@ -970,6 +971,7 @@ export function foo([[a]], used) {
           messageId: 'unusedWithIgnorePattern',
           data: {
             name: 'foo',
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             type: 'Import',
             pattern: DEFAULT_IGNORED_REGEX,
           },
@@ -1044,6 +1046,7 @@ console.log(named);
           messageId: 'unusedWithIgnorePattern',
           data: {
             name: 'defaultImp',
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             type: 'Import',
             pattern: DEFAULT_IGNORED_REGEX,
           },
@@ -1063,6 +1066,7 @@ console.log(named);
           messageId: 'unusedWithIgnorePattern',
           data: {
             name: 'defaultImp',
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             type: 'Import',
             pattern: DEFAULT_IGNORED_REGEX,
           },
@@ -1082,6 +1086,7 @@ console.log(defaultImp);
           messageId: 'unusedWithIgnorePattern',
           data: {
             name: 'named',
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             type: 'Import',
             pattern: DEFAULT_IGNORED_REGEX,
           },
@@ -1101,6 +1106,7 @@ console.log(defaultImp);
           messageId: 'unusedWithIgnorePattern',
           data: {
             name: 'named',
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             type: 'Import',
             pattern: DEFAULT_IGNORED_REGEX,
           },
@@ -1120,6 +1126,7 @@ console.log(named1);
           messageId: 'unusedWithIgnorePattern',
           data: {
             name: 'named2',
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             type: 'Import',
             pattern: DEFAULT_IGNORED_REGEX,
           },